+++ cmeel/consts.py
@@ -1,8 +1,7 @@
 import os
 import sys
 
+CMEEL_PREFIX = "cmeel.prefix"
 SITELIB = os.sep.join(
     ["lib", "python" + ".".join(sys.version.split(".")[:2]), "site-packages"]
 )
-CMEEL_PREFIX = "cmeel.prefix"
-CMEEL_PTH = os.sep.join([CMEEL_PREFIX, SITELIB])