+++ cmeel/pth.py
@@ -1,6 +1,6 @@
-import os
 import sys
+from pathlib import Path
 
-from . import consts
+from .consts import CMEEL_PREFIX, SITELIB
 
-sys.path.append(os.sep.join([sys.path[-1], consts.CMEEL_PTH]))
+sys.path.append(str(Path(__file__).parent.parent / CMEEL_PREFIX / SITELIB))