+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- fix pth when the current prefix is not the last
+
 ## [v0.5.1] - 2022-05-02
 
 - add metadata